Increase ido auto-merge by an order of magnitude.
The default time is 0.7. This is pretty low. It generally takes more
than 0.7 seconds for me to notice ido is asking me to confirm a new
file. I've increased it to 7 seconds as a initial stab at finding
something that doesn't keep me waiting too long if I've mistyped
something, but gives me enough time to confirm new files.